- Renaming all react.js files into jsx - Update webpack config - remove AnimationMixin
