+++ app/app.js
@@ -1,6 +1,6 @@
 import universalRenderer from './utils/UniversalRenderer';
 import alt from 'altInstance';
-import routes from 'routes.js';
+import routes from 'routes.jsx';
 import html from 'base.html';
 
 export default universalRenderer(alt, routes, html);
