+++ app/components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { RouteHandler } from 'react-router';
-import Navigation from 'components/Navigation.react';
+import Navigation from 'components/Navigation';
 
 import 'scss/main';
