+++ app/components/Vote.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Immutable from 'immutable';
 
-import EntryBox from 'components/EntryBox.react';
-import MainSection from 'components/MainSection.react';
-import Scoreboard from 'components/Scoreboard.react';
+import EntryBox from 'components/EntryBox';
+import MainSection from 'components/MainSection';
+import Scoreboard from 'components/Scoreboard';
 import TopicStore from 'stores/TopicStore';
 
 import styles from 'scss/components/_vote';
