+++ webpack/webpack-hot-dev-server.config.js
@@ -14,12 +14,7 @@ var commonLoaders = [
      * TC39 categorises proposals for babel in 4 stages
      * Read more http://babeljs.io/docs/usage/experimental/
      */
-    test: /\.jsx$/,
-    loaders: ["react-hot", "babel-loader?stage=0"],
-    include: path.join(__dirname, "..", "app")
-  },
-  {
-    test: /\.js$/,
+    test: /\.js$|\.jsx$/,
     loaders: ["react-hot", "babel-loader?stage=0"],
     include: path.join(__dirname, "..", "app")
   },
@@ -96,7 +91,7 @@ module.exports = [
     context: path.join(__dirname, "..", "app"),
     entry: {
       app: "./app",
-      header: "./elements/Header.react"
+      header: "./elements/Header"
     },
     target: "node",
     output: {
